chore: stops bundling lit-html
BREAKING CHANGE: Stops bundling lit-html so applications can deduplicate the dependency in their build process. This removes the ability to use the package *directly* in the browser without a build step. To continue doing so, use import maps to map `lit-html` to a CDN URL to address this issue (see documentation for instructions).