refactor: simplifies render queue
Changes how the render queue works. It now uses a simple counter that is incremented for each queued update which is decremeted once the update is done. Only once the last update brings the counter to `0` will `#renderIfIdle` trigger a render. This changes the render to happen after the last update instead of after the each update.